+++ src/components/badge/badge-interface.ts
@@ -1,3 +1,7 @@
+import { WcsSize } from "../../shared-types";
+
 export type BadgeShape = 'normal' | 'rounded';
 
 export type BadgeColor = 'initial' | 'lighter';
+
+export type BadgeSize =  Extract<WcsSize, 'l' | 'm' | 's'>