+++ CHANGELOG.md
@@ -23,6 +23,7 @@ To migrate, please refer to the storybook documentation (v4 to v5).
 - **modal**: add the css variable `--wcs-modal-overflow-y` to control the overflow-y of the modal content
 - **progress-bar**: add the size attribute with available values `m (default)`, `s`
 - **dropdown**: add the size attribute with available values `l`, `m (default)`, `s`
+- **badge**: add the size attribute with available values `l`, `m (default)`, `s`
 - **counter**: add the `disabled` property
 
 ### Changed